# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "contact": {}
    },
    "paths": {
        "/prod/search": {
            "get": {
                "description": "prod INDEX SEARCH",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "prod INDEX API"
                ],
                "summary": "prod INDEX SEARCH API",
                "parameters": [
                    {
                        "type": "string",
                        "description": "상품명 키워드 삼성 노트북",
                        "name": "keyword",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "최저 가격 5000",
                        "name": "minPrice",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "최대 가격 6000",
                        "name": "maxPrice",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "정렬 1|1",
                        "name": "sort",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "시작 번호",
                        "name": "start",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "결과 개수",
                        "name": "limit",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "response.Response": {
            "type": "object",
            "properties": {
                "count": {
                    "type": "integer"
                },
                "desc": {
                    "type": "string",
                    "example": "null"
                },
                "searchList": {},
                "status": {
                    "type": "integer",
                    "example": 200
                }
            }
        }
    }
}